Make FileExtraFields into a private case class of FileEvent
